flatten/refactor a bit more
