+++ SFGServer/Contracts/Requests/GuestRequest.cs
@@ -1,0 +1,3 @@
+namespace SFGServer.Contracts.Requests;
+
+public record struct GuestRequest(string Username);
