+++ SFGServer/Models/EnergyInfoModel.cs
@@ -1,3 +1,3 @@
-﻿namespace SFGServer.Models.Energy;
+﻿namespace SFGServer.Models;
 
 public record struct EnergyInfoModel(int Energy, int MaxEnergy, int EnergyRegenerationRateMs, int LastEnergyAmount, long TimeEnergyWasAtAmount);
