+++ SFGServer/Models/WorldModel.cs
@@ -1,3 +1,3 @@
-﻿namespace SFGServer.Models.World;
+﻿namespace SFGServer.Models;
 
 public record struct WorldModel(string Type, Guid Id, string Name, int PlayerCount);
